fix(storage-provider-node): clean up the storage provider node
