+++ cli/polka-storage-provider/Cargo.toml
@@ -20,7 +20,6 @@ sealed = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
 subxt = { workspace = true }
-subxt-signer = { workspace = true }
 tempfile = { workspace = true }
 thiserror = { workspace = true }
 tokio = { workspace = true }
