+++ cli/polka-storage-provider/src/rpc/server.rs
@@ -18,10 +18,10 @@ use tokio::sync::oneshot::Receiver;
 use tracing::{info, instrument};
 
 use super::{
-    methods::{common::InfoRequest, register_async, wallet::WalletRequest},
+    methods::{common::InfoRequest, register_async},
     version::V0,
 };
-use crate::{substrate, CliError};
+use crate::CliError;
 
 /// Default address to bind the RPC server to.
 pub const RPC_SERVER_DEFAULT_BIND_ADDR: &str = "127.0.0.1:8000";
@@ -29,7 +29,9 @@ pub const RPC_SERVER_DEFAULT_BIND_ADDR: &str = "127.0.0.1:8000";
 /// RPC server shared state.
 pub struct RpcServerState {
     pub start_time: chrono::DateTime<Utc>,
-    pub substrate_client: substrate::Client,
+    // NOTE: this will be "added back" in #388
+    // but instead of a generic client, its storagext
+    // pub substrate_client: substrate::Client,
 }
 
 /// Start the RPC server.
@@ -66,7 +68,6 @@ pub fn create_module(state: Arc<RpcServerState>) -> RpcModule<RpcServerState> {
     let mut module = RpcModule::from_arc(state);
 
     register_async::<InfoRequest, V0>(&mut module);
-    register_async::<WalletRequest, V0>(&mut module);
 
     module
 }
